requirements.txt
insightface>=0.7.3
onnxruntime>=1.16.0
faiss-cpu>=1.7.4
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
